Fix driver hang in csrRoamRoamingStateDisassocRspProcessor
During roam, we disassociate with the current AP before reassociating
with the new one. Upon receiving the disassoc response from PE, we
look-up for the hand-off candidate in the scan results list to get
the Bss description. If this look-up fails, the current code is
unable to recover and leads to a driver hang. While it removes the
roam command from SME active queue, it does not reschedule another
one from the pending queue due to which no pending SME commands
get processed and the SME queue gets stuck.

The recovery mechanism is being fixed as follows:
1. Look-up for the hand-off candidate in the scan results list.
2. If the look-up fails, then we need to disconnect and clean-up.
a) For this, inform HDD about the roam failure and disconnect.
b) Post a disassoc request to PE with BSSID of hand-off candidate
so that it cleans-up the FT session that was created when pre-auth
was completed.
c) Upon receiving the disassoc response, call csrRoamComplete with
reason as eCsrJoinFailure so that SME layer cleans-up as well. This
also takes care of removing the roam command from the active queue
and schedules the next command from the pending queue, thereby
unblocking the SME layer.