Populate Correct Scan Mode for Roam
The init scan request in Riva is failing with reason code
"QWLANFW_STATUS_BMPS_NO_REPLY_TO_NULLFRAME_FROM_AP".  In Riva in case
of roaming, we should be ignoring the tx complete status and hence not
returning a failure when transmission of NULL frame to the AP
fails. However since the scan mode populated in the init scan request
sent from the host is incorrect we are seeing a failure. This fix is
to populate the scan mode in LIM to reflect a roaming type regardless
of the next state.