sched/rt: Do not account zero delta_exec in update_curr_rt()
There are several places of consecutive calls of dequeue_task_rt()
and put_prev_task_rt() in the scheduler. For example, function
rt_mutex_setprio() does it.

The both calls lead to update_curr_rt(), the second of it receives
zeroed delta_exec. The only effective action in this case is call of
sched_rt_avg_update(), which can change rq->age_stamp and rq->rt_avg.
But it is possible in case of ""floating"" rq->clock. This fact is
not reasonable to be accounted. Another actions do nothing.

Signed-off-by: Kirill V Tkhai <tkhai@yandex.ru>
CC: Steven Rostedt <rostedt@goodmis.org>
CC: Ingo Molnar <mingo@kernel.org>
CC: Peter Zijlstra <peterz@infradead.org>
CC: linux-rt-users <linux-rt-users@vger.kernel.org>